Align content width with media switch
When we're in 'laptop mode', but screen width is smaller than
content width, there appears a horizontal scrollbar and a couple
of artifacts that look pretty awful. This change avoids this
by reducing contents width.

We also return the use of variables for calculating the actual
width of content, and do it in SaSS, instead of browser side.