solution: Rewrite links layout
The whole link-inside-div-inside-table-inside-link idea
was just a bit too much. Instead, this change builds
a list of links, which is a good mental model and
looks awesome in text browsers.

Adjusting baselines of spans using css as we do is
a bit daring, but seems to work just fine.

The link boxes became a bit wider, which looks ok to me.
We can achieve pixel-perfect match with previous version
by playing with padding values, but I don't think we
need to.